Function on Marshmallow
- Version runs on Android 6.0; avatar scaling for different display
sizes corrected
- Journey and Event header pictures changed to match the chosen
journey/event
- Map added to EventView

+++ app/src/main/java/zollie/travelblogger/guidee/utils/ImageProcessor.java
@@ -45,7 +45,7 @@ public class ImageProcessor {
 
         return circleBitmap;
     }
-    public Bitmap resizeMarkerImage(Bitmap myPic, int size){
+    public Bitmap resizeMarkerImage(Bitmap myPic, float size){
 
         Matrix m = new Matrix();
         m.setRectToRect(new RectF(0, 0, myPic.getWidth(), myPic.getHeight()), new RectF(0, 0, 50*size, 50*size), Matrix.ScaleToFit.CENTER);
